+++ src/Checkbox/index.tsx
@@ -1,3 +1,5 @@
 import { Checkbox as AntdCheckbox } from 'antd';
 
+export { CheckboxProps } from 'antd';
+
 export const Checkbox: typeof AntdCheckbox = AntdCheckbox;