+++ src/Radio/index.tsx
@@ -1,3 +1,5 @@
 import { Radio as AntdRadio } from 'antd';
 
+export { RadioProps } from 'antd';
+
 export const Radio: typeof AntdRadio = AntdRadio;