+++ src/Timeline/index.tsx
@@ -1,7 +1,9 @@
 import {
   Timeline as AntdTimeline,
   TimelineProps as AntdTimelineProps,
+  TimelineItemProps as AntdTimelineItemProps,
 } from 'antd';
 
 export const Timeline: typeof AntdTimeline = AntdTimeline;
 export type TimelineProps = AntdTimelineProps;
+export type TimelineItemProps = AntdTimelineItemProps;